DataclassToProto: Fix bad mutable default in sample lib training block
This is an interesting one where using a dataclass actually caught a bug!
It seems that dataclass refuses to use a mutable default as the default
value for a field in a dataclass and raises an error telling you to use a
default_factory instead. This led to the place where we're auto-deducing a
dataclass from the args in this train function and thus to the bad mutable
default.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>