DataObjectBase: Remove all of the automagic logic for matching data objects to precompiled protos
This logic was responsible for auto-deducing a precompiled protobuf class
that matched the name of the data object class. Since we're moving all of
the precompiled protobuf logic to being a "power user feature," it's fair
to require that any data objects following this pattern explicitly declare
it by setting the _proto_class attribute explicitly.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>