TrainBackends: Refactor the backend section of caikit.config to split load/train
This also includes some significant reframing in how backends are
configured. Rather than a separate mapping of configs and a list of
priorities, there is a single priority list where the entries hold the type
of the backend as well as the config. This format also allows for duplicate
entries for a given type that are disambiguated by a "name" key to allow
for multiple instances of the same backend.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>