TrainBackends: Add the ability for a module to know its load backend
This will allow backend-specific module implementations to fetch the
specific instance of the backend that was used to load them. This, then,
allows multiple named instances of a given backend type to coexist as long
as the backend type can successfully "pass" on loading a given module that
is not right for the named instance.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>