TrainBackends: Overhaul module backend config for train/load split
There are now two separate global registries and all lookups are done
explicitly for load/train. The registries are also now lists rather than
a dict. While dicts are strictly ordered since 3.8, it's not really the
right data structure to hold ordered information. Since backends need to
be held in priority order, a list is a better representation.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>