TrainBackends: Overhaul model manager to support shared loaders
This majorly reframes how the backend loading is done. There are a couple
of critical changes:

1. Backends are always tried in priority order
2. Shared loaders can be tried before any attempt is made to parse a
config.yml, so special loaders can be built that don't require one

The functionality to match up a module's artifacts with supported load
backends is preserved for now, but may be worth removing eventually.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>