NumpyTypeSupport: Add numpy as a dependency
This one is up for discussion! This does break the rule that caikit itself
will not hold any heavy (relatively) mathematical libraries. On the other
hand, I don't know of a single AI library that _doesn't_ depend on numpy
and bringing it in here allows for the dtype framework to be used for more
granular type mapping in the data model

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>