FullOneofSupport: Update getattr semantics to support getting oneofs by name or field
Oneof values will always be stored in python under their oneof name (TODO
for precompiled protos), so this commit adds logic to look up that value by
either the name of the oneof itself or the name of the field. The field
will only return if it is the correctly set field.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>