FullOneofSupport: Add setattr semantics for oneof fields
For oneof field names, we alias to the oneof name itself to store the value
and rely on the inverse semantics at getattr time.

Still TODO! If the oneof name itself is used to set the oneof, we need to
determine which of the oneof fields is being set. This may require mapping
from python types to proto types, so I'd like to try to find a more elegant
solution.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>