FullOneofSupport: Add oneof field support to default __init__ semantics
When no custom __init__ is provided, this allows the default __init__ to
take the names of the oneof fields and alias them to the oneof itself while
correctly setting the which_oneof value. This partially solves the setattr
issue by providing a way to explicitly set which one is being used by using
the specific field name.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>