FullOneofSupport: Shore up oneof semantics for precompiled protobufs
The crux of this change is that for a class that pre-defines _proto_class
and does not rely on @dataobject to inject it, we need to shuffle the
predefined set of fields to include the oneof names and NOT include the
names of the oneof fields. This emulates how data is stored and accessed in
a dataobject.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>