Changes for hosting API docs on readthedocs
readthedocs requires the docs to be generated from  source code
automatically during readthedocs build as it runs
'sphinx-build' only and not 'sphinx-apidoc' as well.

This PR adds autoapi extension to the sphinx config to run
the automatic generation of docs from the source files. It also
tidies the sphinx files into 'source' directory and removes the
redundant call to 'sphinx-apidoc'.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>