Remove the module_id warnings from unit tests
Some of the test models - those defined in zip files, where not
updated to the new config format with module_id, which caused a
lot of warnings being logged when running unit tests.

Update the zipped models to use the module_id instead. Add a new
test that specifically covers the no-module-id deprecation warning,
verifying that the deprecation is raised via pytest.deprecated_call()

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>