Add e2e test for examples
Add an e2e test for the text sentiment example.
The test is written as a standard python test so it can be
discovered by pytest and all the nice python tooling for
tests works as usual.

The test creates a new virtual environment and installs
caikit from the local disk, to ensure it tests the version of
caikit from the pull request (or local code changes).

The test then follows the same process described in the README:
- installs requirements.txt
- runs the start_runtime.py script
- runs the client.py

This test ensures that the example stays in a working state.
The example is likely to be one of the first points of contact
of people looking at caikit and it should always be working.

This test also helps in detecting unintended API changes.
Eventually we should have a API driven test suite that covers
the whole API.

The new test is not executed by default via tox, to keep
tests execution on laptops as quick as possible. The new
test is enabled in the GitHub action workflow.

Fixes: #151

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>

Rewrite test_examples to be more reusable

Move reusable logic to a shared module with examples.
Use module level test_ function instead of a test class to be
more consistent with other tests in caikit.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>