Fix example and test
Caikit needs to be installed with "pip install caikit[all]".
Add "[all]" to the example requirements.txt to fix the
example and to the examples test as well.

The change in the requirements configuration was merged
before the example test was, so main is now broken.

This change also remove the delelopment mode from the
pip command in the test examples, since it does not work
without a setup.py from python3.9 and earlier.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>