ReorganizeModuleBackends: loaders -> initializers
We talked a lot about this and the word "loader" is too overloaded in the
ML world. The confusion stemmed from thinking it was "the thing that
unpacks matrix weights into memory" versus "the thing that brings a black
box class into memory for use." After a lot of discussion, we decided on
initializer as the best non-confusing noun.

https://github.com/caikit/caikit/issues/54

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>