TrainFutures: Make save_with_id a central part of the Trainer.train interface
Every trainer needs the ability to own its own set of IDs so that they can
be retrieved independent of the in-memory state of the the instance of
caikit that is running. As such, the job of allocating ids needs to be left
with the Trainer, but we also want some central control over how models are
saved, so we want the inputs to the final path to all be arguments to the
train function itself. The cleanest way to do this is to have central logic
in the base class that children can share to inject the id as needed.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>