TrainFutures: Add instance_name to generic factory
Factory constructibles must be able to take an instance name. They're free
to ignore it, but it will be passed through construct so that they have
access to it if it's known. This is needed for factory config secitons that
are managed as key:val maps where the val blobs may not be unique w.r.t.
the name of the class.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>