TrainFutures: First pass overhaul of runtime to use MODEL_MANAGER for training
This has some pretty significant changes, most notably that auto-loading
trained models is gone. This configuration was only ever used for testing,
and auto-loading without a state management engine is not robust to multi-
replica deployments of runtime, so it was going to need an overhaul before
ever being used in a production environment.

Left TODO after this commit is the rest of the test cleanup!

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>