TrainFutures: Add option to LocalModelTrainer to use a subprocess
I opted to have a single trainer with a toggle rather than two separate
trainers because all of the wrapper logic beyond the execution containment
tool was shared. We could have split it all out into a base class, but this
is a bit simpler code-wise and doesn't limit potential users in any
meaningful way.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

# Conflicts:
# caikit/config/config.yml