FlakeyWipTest: Convert tests to pytest and loosen regex match checks
I'm pretty sure that these sporatic errors are caused by the introduction
of pytest-asyncio as a test plugin. The tests themselves were explicitly
checking caplog.records[0], so if some other warning pops up during the
capture, the regex check was failing. Since the consistent error has to do
with the asyncio pool being closed, I suspect this is just a bad
coincidence plus a test that's assuming it's more isolated than it is.

https://github.com/caikit/caikit/issues/305
Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>