LocalTrainer: Handle threaded workers that can't be interrupted
The DestroyableThread is not able to interrupt threads that are busy in a
C extension that is not controlled by the interpreter, so it's very likely
that destroying a training job may not immediately cause it to stop when
using threading. To handle this, we don't need to wait for the worker to
join on cancel(), we only need to ensure that the worker has been told to
cancel.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>