LazyLocalModelLoading: Fix logic for when local_models_dir sync gets initialized
We want to allow a default config value for local_models_dir without
requiring that it's enabled, so if the dir does not exist, the whole thing
is considered disabled. We then want to kick of synchronization IFF the
local_models_dir exists and not condition it on the presence of any models
since it is no longer just a boot-time process.

This also means that it no longer raises if the initial sync does not load
any models.

https://github.com/caikit/caikit/issues/324

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>