LazyLocalModelLoading: Make sure loading has finished before unloading
This could potentially waste some resources, but it ensures that there
won't be any additional data brought into memory from orphaned loading
after the model is no longer in the map.

We could explore using an AbortableAction here with an alterante
ActionAborter implementation, but that's likely not worth the added
complexity.

https://github.com/caikit/caikit/issues/324

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>