UseSpawnForTrainSubprocess: Wrap up stream sources for safe pickling in servicer_util
This allows DataStream types in function signatures that get translated to
DataStreamSource types in the handlers to be passed to subprocesses using
pickle safely, even when using "spawn." The key is to recreate the dynamic
stream source type before trying to actually unpickle it. Thiw will be slow
relative to an inference call, so should be used carefully there, but this
is unlikely to be used there anyway, and it should be a small amount of
overhead relative to a training job.

https://github.com/caikit/caikit/issues/384

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>