HealthProbeCommand: Full tests for all health probe configurations
Also, a bunch of bug fixes! These fixes attempt to be as secure as
possible with only the TLS credentials available through the config.
This means that we need to not verify the server cert as well since we
don't know for sure that we'll have the server's CA (it may not be
self-signed). In the context of a health probe, this is ok since it is
expected to run inside of the same trusted bondary as the server itself.

https://github.com/caikit/caikit/issues/548

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>