HealthProbeCommand: Move the health probe to its own package
Importing caikit is slow (~300ms on my M1 Mac), so having the health probe
live inside of caikit.runtime was adding an unnecessary 300ms to the probe
latency. The only portion of caikit that needs to be shared is the unified
config, so having it in its own package (and carefully avoiding importing
caikit when importing caikit.config) allows the probe to run without the
overhead of the full caikit import.

For context, the majority of the import latency comes from the core's data
model which does a bunch of work at import time to set up the common data
objects used in the core.

https://github.com/caikit/caikit/issues/548

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>