ModelFetchOutOfAsyncio: Move the model lookup for inference signature into predict_model
The `retrieve_model` operation can potentially be expensive. With the
`async` HTTP server, doing this outside of the `run_in_executor` call
meant that the work to look up the model (and potentially wait for it to
finish loading) was blocking the event loop. That work to fetch the
model was _only_ being done so that the inference function could be
deduced, so moving that logic inside of `predict_model` allows the
lookup (and load) logic to not block the event loop.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>