ModelService: Add deploy_model / undeploy_model to runtime model manager
These implementations rely on having write access to local_models_dir and
treating that as the central location for models to be deployed/undeployed.
A more robust implementation would involve abstracting out all the use of
local_models_dir into a ModelStorage type and letting the local-disk impl
stand as-is. Given our pullback from this type of abstraction, we're going
to stick with hard-coded disk as the medium for now.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>