otel: Fully remove the internal request_id
One of the big follow ups I'd like to do from this PR is to create an
actual RuntimeServerContextType class that does the shimming between grpc
and fastapi so that we don't need to keep writing optional dependency
checks and type checks. If we do this, I think this would be a fine place
to reinstate this request_id logic so that the id becomes available to any
place downstream that wants to have it for logging purposes.

As it stands, though, the only use of this request_id is to attach it to
the trace which is useless because it will always have a unique ID for the
span anyway.

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>